requirements.txt
Django==1.5.4
MySQL-python==1.2.4
South==0.8.2
django-nose==1.2
django-tastypie==0.10.0
factory-boy==2.2.1
fuzzywuzzy==0.2
lettuce==0.2.19
nose==1.3.0
python-dateutil==2.1
python-mimeparse==0.1.4
selenium==2.37.2
six==1.4.1
splinter==0.5.4
sure==1.2.2
wsgiref==0.1.2
